move base_test to test_helpers
