+++ cli/command/test_helpers.go
@@ -4,9 +4,39 @@ import (
 	"flag"
 	"testing"
 
+	"github.com/golang/mock/gomock"
+	"github.com/quintilesims/layer0/cli/printer"
+	"github.com/quintilesims/layer0/cli/resolver/mock_resolver"
+	"github.com/quintilesims/layer0/client/mock_client"
 	"github.com/quintilesims/layer0/common/config"
 	"github.com/urfave/cli"
 )
+
+type TestCommandBase struct {
+	client   *mock_client.MockClient
+	printer  *printer.TestPrinter
+	resolver *mock_resolver.MockResolver
+}
+
+func newTestCommand(t *testing.T) (*TestCommandBase, *gomock.Controller) {
+	ctrl := gomock.NewController(t)
+
+	tc := &TestCommandBase{
+		client:   mock_client.NewMockClient(ctrl),
+		printer:  &printer.TestPrinter{},
+		resolver: mock_resolver.NewMockResolver(ctrl),
+	}
+
+	return tc, ctrl
+}
+
+func (tc *TestCommandBase) Command() *CommandBase {
+	return &CommandBase{
+		client:   tc.client,
+		printer:  tc.printer,
+		resolver: tc.resolver,
+	}
+}
 
 func getCLIContext(t *testing.T, args []string, flags map[string]interface{}) *cli.Context {
 	flagSet := &flag.FlagSet{}
